fix unitests

+++ python/paddle/fluid/tests/unittests/test_activation_op.py
@@ -168,7 +168,7 @@ class TestParameter:
     def test_out_name(self):
         with fluid.program_guard(fluid.Program()):
             np_x = np.array([0.1])
-            data = paddle.static.data(name="X", shape=[-1, 1])
+            data = paddle.static.data(name="X", shape=[-1, 1], dtype="float32")
             out = eval("paddle.%s(data, name='Y')" % self.op_type)
             place = fluid.CPUPlace()
             exe = fluid.Executor(place)
@@ -521,7 +521,7 @@ class TestAtan(TestActivation, TestParameter):
     def test_out_name(self):
         with fluid.program_guard(fluid.Program()):
             np_x = np.array([0.1])
-            data = paddle.static.data(name="X", shape=[-1, 1])
+            data = paddle.static.data(name="X", shape=[-1, 1], dtype="float32")
             out = paddle.atan(data, name='Y')
             place = fluid.CPUPlace()
             exe = fluid.Executor(place)
